requirements.txt
streamlit
Pillow
numpy
scikit-learn
scikit-image
opencv-python-headless
joblib
